requirements.txt
selene == 2.0.0rc2
pytest-voluptuous == 1.2.0

pytest == 7.2.0
voluptuous == 0.13.1
python-dotenv == 0.21.0
allure-pytest == 2.12.0
curlify == 2.2.1
requests == 2.31.0
curlify~=2.2.1

jsonschema~=4.17.3
Faker == ^18.13.0